WIP: Big mess of rework to filters.
* Filters now operate on the entire 'package' rather than just the
  'killmail' sub-section.  Needed for:
* minimum_value filter which reads package['zkb']['totalValue']

* 'exclude_if_any' inverts the usual sense of filters: we'll skip
  a killmail if it matches one of the filters in this area.  Useful
  to prevent the bot from posting about your own losses, or about kills
  happening 'in chain' that are happening in Jita/Amarr which perhaps are
  pinned to your Siggy.

* Filters working as before are now in 'require_all_of'.
  Perhaps this name is confusing and could be better?
  Or perhaps we can fix it by making the filter names a bit better --
  e.g. "location_is_in" or "location_one_of"

I want to change the idea of filters (returns True if we should 'skip' a mail)
into the notion of matchers (returns True if a mail matches a given property)
just because that's much easier to reason about, especially given the new
require/exclude distinction.

I also want to change the set of filters away from being a dict into being
a list of dicts, so you can do e.g.
                "require_all_of": [
                    {"location": ["chain"]},
                    {"location": ["wspace"]}
                ]