add --oneshot-killmail-file for testing
--oneshot-killmail-file takes a JSON file and, after initializing
mappers as usual, runs it through one iteration of the usual loop
and then exits.  useful for testing new configs, for debugging, and
for hacking on pretty killmail formatting.

as part of doing this, break out initialization of the main loop
into a helper function _initialize().  and as part of doing that,
run a loop of the mappers before initializing subscribers (e.g. kills).
this fixes #34